requirements.txt
gunicorn==20.1.0
Flask==2.0.1
MarkupSafe==2.0.1
redis==3.5.3
cryptography==3.4.7
Flask_S3==0.3.3
python-environ==0.4.54